+++ app/controllers/GQLSchemaController.scala
@@ -1,8 +1,8 @@
 package controllers
 
 import javax.inject.{Inject, Singleton}
-import models.GQLSchema
-import play.api.mvc.{AbstractController, ControllerComponents}
+import models.gql.GQLSchema
+import play.api.mvc.{AbstractController, Action, AnyContent, ControllerComponents}
 import sangria.renderer.SchemaRenderer
 
 import scala.concurrent.ExecutionContext
@@ -11,12 +11,12 @@ import scala.concurrent.ExecutionContext
 class GQLSchemaController @Inject()(implicit ec: ExecutionContext, cc: ControllerComponents)
   extends AbstractController(cc) {
 
-  def renderSchema =
+  def renderSchema: Action[AnyContent] =
     Action {
       Ok(SchemaRenderer.renderSchema(GQLSchema.schema))
     }
 
-  def renderClient =
+  def renderClient: Action[AnyContent] =
     Action {
       Ok(views.html.graphiql(None))
     }
